pyproject.toml
[project]
name = "whisper-sae"
version = "0.1.0"
description = "Sparse Autoencoder Analysis for Whisper Speech Model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "sae-lens>=4.0.0",
    "einops>=0.7.0",
    "wandb>=0.16.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
    "jiwer>=3.0.0",
    "scipy>=1.11.0",
    "statsmodels>=0.14.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "soundfile>=0.13.1",
    "librosa>=0.11.0",
    "ipywidgets>=8.1.8",
    "circuitsvis>=1.43.3",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "ipykernel>=6.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]
modal = [
    "modal>=0.56.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
